Minor static fixes, precheck some variables, build local fix-ca.
Many functions are for internal use, added static to avoid warning.

Reduced the nesting in the switch by prechecking necessay variables
and exiting early.

In Issue 1 it was noted that we can't build fix-ca. Since fix-ca is
just one file, it seemed we can still continue to use gimp-tool-2.0
to build and install the file like before if we avoid including the
fix-ca-config.h - this can be useful for users who need this filter
but do not have admin access.