gimp_preview_draw_buffer() works differently from prior preview mode
Moving the preview window to the right discovered this new window is
[0,0,width,height] based, therefore I had to add a hack to work with
original image dimensions plus then copy image to location [0,0] and
size width,height.

There also appears to be a problem with fix_ca_region() working with
a smaller image of size [width,height], therefore the need for using
the full-sized image.

Fixed round_nearest() to work with rounding negative values.