Include another Lateral CA example, and update README.md
README.md moved the "using" info to the top since expect
most readers are Fix-CA "users" and moved the "building"
and compile to the bottom since this is only of interest
for those that are building Fix-CA.
Added new info for lens center with example branches.jpg

Added "half" testing to configure.ac which can be useful
for adding RGB _Float16 later. Basically a future TODO.

Checked and fixed some gimp_install_procedure() details.