Ensure we have enough pixel range for correcting very large pictures.
Finding that I had to move blue +6 pixels to correct example branches
which is a small image by todays standards, it made sense to increase
the +/- range we can move pixels. Most pictures are only likely going
to need moving +/-one or two pixels but having the extra range helps.